fixed training loop, but syncing changes with origin branch broke the new eval loop. working on that now. I have a feeling it's going to require full integration of channels first.

+++ torch_mesmer/mesmer.py
@@ -204,7 +204,7 @@ class Mesmer():
                                         nuclear_kwargs = postprocess_kwargs_nuclear
                                         )
 
-        label_image = resize_output(label_image, orig_img_shape)
+        label_image = resize_output(label_image, orig_img_shape).astype(int)
 
         if return_transforms:
             return label_image, output_images
